LayoutTests/fast/dom/Range/range-detach-noop.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests that Range.detach() is a no-op.");

var range = document.createRange();
shouldNotThrow("range.detach()");
shouldNotThrow("range.detach()");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
